+++ test/libsolidity/ABIJson/empty_name_input_parameter_with_named_one.sol
@@ -13,10 +13,12 @@ contract test {
 //     "inputs":
 //     [
 //       {
+//         "internalType": "uint256",
 //         "name": "",
 //         "type": "uint256"
 //       },
 //       {
+//         "internalType": "uint256",
 //         "name": "k",
 //         "type": "uint256"
 //       }
@@ -25,10 +27,12 @@ contract test {
 //     "outputs":
 //     [
 //       {
+//         "internalType": "uint256",
 //         "name": "ret_k",
 //         "type": "uint256"
 //       },
 //       {
+//         "internalType": "uint256",
 //         "name": "ret_g",
 //         "type": "uint256"
 //       }