+++ test/libsolidity/ABIJson/return_structs.sol
@@ -15,6 +15,7 @@ contract C {
 //     "outputs":
 //     [
 //       {
+//         "internalType": "uint256",
 //         "name": "x",
 //         "type": "uint256"
 //       },
@@ -22,6 +23,7 @@ contract C {
 //         "components":
 //         [
 //           {
+//             "internalType": "uint256",
 //             "name": "a",
 //             "type": "uint256"
 //           },
@@ -29,14 +31,17 @@ contract C {
 //             "components":
 //             [
 //               {
+//                 "internalType": "uint256[2]",
 //                 "name": "x",
 //                 "type": "uint256[2]"
 //               }
 //             ],
+//             "internalType": "struct C.T[]",
 //             "name": "sub",
 //             "type": "tuple[]"
 //           }
 //         ],
+//         "internalType": "struct C.S",
 //         "name": "s",
 //         "type": "tuple"
 //       }