+++ Changelog.md
@@ -11,6 +11,7 @@ Compiler Features:
  * eWasm: Highly experimental eWasm output using ``--ewasm`` in the commandline interface or output selection of ``ewasm.wast`` in standard-json.
  * Metadata: Update the swarm hash, changes ``bzzr0`` to ``bzzr1`` and urls to use ``bzz-raw://``.
  * Standard JSON Interface: Compile only selected sources and contracts.
+ * Standard JSON Interface: Provide secondary error locations (e.g. the source position of other conflicting declarations).