+++ Changelog.md
@@ -19,6 +19,7 @@ Compiler Features:
 Bugfixes:
  * View/Pure Checker: Properly detect state variable access through base class.
  * Yul analyzer: Check availability of data objects already in analysis phase.
+ * Yul Optimizer: Fix an issue where memory-accessing code was removed even though ``msize`` was used in the program.